+++ lib/feedback.dart
@@ -2,3 +2,5 @@ library feedback;
 
 export 'src/better_feedback.dart';
 export 'src/feedback_functions.dart';
+export 'src/l18n/translation.dart';
+export 'src/theme/feedback_theme.dart' show FeedbackThemeData;